Implemented HTML5-templates to display contacts
This is much better than creating strings with raw HTML and modifying
elem.innerHTML, both performance-wise and for the sake of readability.

Together with this change I also improved performance of the function
updating the contactsList, since it now only will add/remove necessary
contacts and not deleting & re-rendering the whole list on every small
change.

localStorage and Template-tags are pretty nice actually!